+++ backend/src/app/model/product.ts
