+++ backend/src/seed/products.json
@@ -114,8 +114,6 @@
     "name": "Aurora Blue Shirt",
     "image": "shirt.png"
   },
-
-
   {
     "gtin": "1234567891123",
     "brandName": "Nova",
